chore: consolidate frontend api helper
